cs50p/week6/shirt/pyproject.toml
[tool.poetry]
name = "shirt"
version = "0.1.0"
description = ""
authors = ["Kostiantyn Gurianov <kguryanov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
